Reverted sim get_state to no keyframe states
Since ground truth is intended to describe the actual state of the MAV
and the simulation itself is only involved with the global state
(keyframe is entirely internal to the filter) it doesn't make sense to
have the sim return truth states for the keyframe. This also makes
implementing the keyframe reset easier.